refactor: Update menu container background color and animation
The menu container background color and animation have been updated in the style.css file. The background color is now set to the primary color variable, and a new animation called "grow_circle_path" has been added to create a growing circle effect. The clip-path property has also been modified to control the size and position of the circle. This change improves the visual appeal of the menu container.